requirements.txt
flask==2.0.1
openai==0.27.0
langchain==0.0.266
unstructured==0.7.11
chromadb==0.3.26
tiktoken==0.4.0
python-dotenv
hdbcli 
sqlalchemy-hana
OpenAI
#langchain openai sqlalchemy sqlalchemy-hana
langchain
openai
langchain_experimental
SpeechRecognition
PyAudio